Put .cdb inside world.json and expose Sheets into FieldInstances
Many things depended on eachother, so this turned into on huge commit.

The .cdb file now lives as a string in the main JSON file. This fixes
the issue of having to carry around an extra file. Exporting to .cdb
will come later

Entities now also have a new datatype "Sheet" (still kind of broken in
this commit)